cmdvartab: update descriptions
while the descriptions are not strictly necessary, in various UI places like
the cgi web pages you don't get the name of the var itself only the description
from the cmdvartab file. if the description isn't in the file then you just get
the default of 'Description unavailable', this is usually repeated lots of
times making the UI useless.
in other places the description is a paragraph and a half which again makes the
UI useless.

add some typical missing ones and shorten the stuff that's way too long