silence common/common.c upsnotify properly
upsnotify: failed to notify about state 2: no notification tech defined, will not spam more about it

syslog gets spammed with this message a lot.
The message is also vaguely misleading as it could easily imply you just need
to define something in a config file somewhere.

Someone complained about it on the mailing list:
https://www.mail-archive.com/nut-upsuser@alioth-lists.debian.net/msg02295.html

however the reply was very unconvincing:

""Technically, it is not about just systemd - other systems where launched
processes that have ways to interact with some OS framework (like SMF,
upstart, Windows services, docker/kubernetes, whatever) *and* such
framework offers ways for services to inform they have started and e.g.
dependencies can begin starting, with better precision than "we have forked
off an init script and assume that this instant its service is fully ready".""

however when you wade through the #ifdef mess in the upsnotify function it
(eventually) becomes clear that it doesn't support any 'notification tech'
other than systemd and that when nut is compiled in the abscense of systemd
the whole purpose of upsnotify is to spam the logs with a message that
effectively says

"you had the audacity to compile on a platform that doesn't have systemd so
 I'm going to complain about it incessantly until you mend your ways"

nut builds for all sorts of systems, most of which are never going to be able
to run the linux-only systemd - windows, *bsd, mac, etc. and even on linux
it's not universal given it's just another user level program.

yet what do we do? penalise everyone else. nice, highly obnoxious behaviour.

you could maybe understand having a notification framework if it was agnostic
but in this case it's not, it's doing things in a systemd specific way.

the better way to have done it would be to use weak functions with a default
implementation that does nothing. you then link in a more specific version
for the platform you're building for, so for example notify-windows.c,
notify-systemd.c, notify-syslog.c, notify-solaris.c and so on. this would
lead to a transparent and properly abstracted interface without the mess of
one thing

I don't have sufficient spherical tuits to do that myself, so just stub out
the function and (hopefully) let the compiler eliminate it for us in a
relatively non-intrusive fashion