upsset.cgi: upsset.c:760: do_type: Assertion `l <= 127' failed.
the code has a comment:
/* FIXME: Loophole about longer numbers? Why are we limited to char at all here? */

the clue is that you're not. you're about to call into do_string with that l param
as a 'char' however do_string declares the param as an int and it's only used to
emit some html. this is also the sole use of do_string so it's not like there's
anything else to be bothered about.

thing is, with the assert in the code upsset.cgi just crashes with the message,
there's no useful information provided to help.

when you remove the silly assert it seems that what's actually causing it is the
following two vars

driver.debug
driver.flag.allow_killpower

both of which have Type: NUMBER, and don't have a Maximum Length defined.
it sort of seems that upsset hasn't been taught how to deal with NUMBER types
so more investigation is likely needed here

That said, there are four other vars in my dump from apc-mib with a length
of 128, so this is still valid - there's absolutely no reason to force it
down to a char just to implicitly convert back to an int