upsset: remove I_HAVE_SECURED_MY_CGI_DIRECTORY
this very much does not do what it intends to, it has no way to determine
whether you have actually secured the cgi directory, or not. certainly it
has no way to enforce it.

all it really does is show how badly designed / implemented it is.
there's effectively no security here at all and no actual attempt at any,
this might have been ok in 1993 but it's now 2023