remove html 4.01
ok, it's 2023, not 1993. html 4 is long dead.
also, sticking those external 'valid' icons on there gives a way for a side
channel attack on what's highly likely to be an internal-only resource.
this raises a potential security vulnerability and shows again the bad design,
yes it may have been ok to do stuff like this in 1993, but the world has moved
on.