apc-mib: add ups date and time
this provides the ups idea of it's internal date/time into the ups.date &
ups.time vars.

according to the PowerNet-MIB the date should be in mm/dd/yy format but all
of the ones I have tested with provide mm/dd/yyyy
dates from the PowerNet-MIB are inconsistent with some being influenced by
the date setting for the 'apc' user in the web UI. due to this inconsistency
you can't really know what format you're going to get and as such you can't
rely on the date or make use of it for much more than as a display item that
allows a human to see if the UPS date matches reality.

at least the NMC2 cards, likely others, allow configuring the date/time to
sync with an ntp server and that's what you should do anyway, but having
access to these from management applications makes it much easier to confirm
that's still happening