upsmon: disable wall by default
wall isn't something you want turned on by default, it makes the config very
verbose as you need to discover all the events and then specifically turn
wall off for every one - and that assumes it's not blindly calling wall
anyway while ignoring the config.

wall is useless on a headless server anyway and that probably covers the
major use-case today - servers in racks doing a job with nobody logged in.

the wall usecase of a desktop with it's own UPS has largely moved on to
people with laptops which have their own battery, so it's simply not
relevant anymore