apc-mib: implement shutdown.reboot
when driver.killpower is called, snmp-ups.c calls

shutdown.return
shutdown.reboot
load.off.delay

in turn, but apc-mib.c only provides shutdown.return which maps to
.1.3.6.1.4.1.318.1.1.1.6.1.1.0 = 2
PowerNet-MIB::upsBasicControlConserveBattery.0 = turnUpsOffToConserveBattery(2)

The description in the MIB itself says

Setting this variable to turnUpsOffToConserveBattery(2) causes a UPS on battery
to be put into 'sleep' mode. The UPS will turn back on when utility power is
restored. Attempting to turn off a UPS that is not on battery will result in a
badValue error. Setting this value to noTurnOffUps(1) has no effect.

So you have a race here, if the power is restored somewhere between you
tripping FSD and calling driver.killpower/shutdown.return then setting this MIB
entry will fail. As there are various configurable delays and ultimately
timeouts between these two points (i.e. HOSTSYNC by default is 30 seconds) then
there is reasonable opportunity for power to return and the command to fail.
How probable this is in reality is hard to estimate, but you can be sure that
if it does happen then it will be at the most inopportune time.

Given neither shutdown.reboot or load.off.delay are implemented then you fail
to turn the UPS output off and all your servers stay off as they never get the
chance to restart due to the power cycle.

Therefore implement shutdown.reboot using
PowerNet-MIB::upsAdvControlRebootShutdownUps = rebootShutdownUps(2)
.1.3.6.1.4.1.318.1.1.1.6.2.2.0 = 2

per the description in the MIB:
Setting this variable to rebootShutdownUps(2) causes the UPS to shut off and
turn back on (as long as good AC input power is available, otherwise the UPS
will remain off until good AC input power is available).

this means we're guaranteed to switch the output off, even if we're no longer
OB.