apc-mib: add further outlet controls
the apc PowerNet-MIB has various outlet specific controls that can be used
to provide more control over various aspects of any switched outlets.

adding most of this with names that aren't really covered in the nut var
name descriptions, but hopefully roughly in the right place

of interest for me is the outlet delayed shutdown reboot command which can
be used to shutdown an outlet after a delay and have it come back when mains
returns. this is due to having a piece of equipment attached to a server
that can't run nut and can't be otherwise remotely powered. so instead just
before the server powers off we can set a time delay for the equipment to be
powered down a short interval after the server to prevent it unnecessarily
draining the UPS batteries further