cgi web pages: display upsstats by default
by default when you go to the web pages you get a blank page with the option
to select stats or settings, so no matter what you're required to make an
extra click. likleyhood is what you actually want is a quick overview of the
status of the multiple UPS you have configured, making that take an extra
click is pointles and annoying.  Yes you require an extra click to get to
settings but that's no worse than at present