apc-mib: flag invalid values
in various instances the NMC will return invalid values. for example it'll
return zero for humidity if there isn't a sensor plugged in.
flag a few of these instances such that they don't get propagated in a
manner that's totally useless. this also makes it easy to use
default.ups.power.nominal in ups.config to replace a nonesense value